More advanced scheduling (#128)
Adds filtering and scoring plugins.
The new behaviour  should be close to how it was before.
Things that will have changed: now takes into account pod requests and decides if a node has capacity if those are set.
When calculating capacity, a default pod request of 1m and 200mb are used if none are set on the pod.
We don't take into account pod overhead.
There's a lot left to do on node resources fit plugin.